data-producer/requirements.txt
kafka-python==2.0.2
PyYAML==5.4.1
